+++ Assets/RoR2Shaders/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.0
+- Added Dithering shader
+- Added Hue Saturation Value shader
+- Added Sharpeness shader
+
 ## 1.2.0
 - Added Color Banding shader